==> 04_psychometry/09_2_p_val_gc_correction_maxh.R <==
# Aim of this script is to make correction for Genomic control (GC) for MaxH

library(data.table)

intercept_data <- read.csv('../../../data/03_neurodegenerative_diseases/several_traits/four_traits/MaxH/intrecept_data.csv') # load table with parameters for GC 
gwas <- fread('../../../data/03_neurodegenerative_diseases/several_traits/four_traits/MaxH/02_unification_results/maxh_output_done.csv')

# Make GC correction for intercept
chi <- qchisq(gwas$p, df = 1, lower.tail = F)
chi_gc <- chi/intercept_data$intercept[1]
gwas$p_gc <- pchisq(chi_gc, df = 1, lower.tail = F)

# Write corrected GWAS summary statistics
fwrite(gwas, row.names=F, file = '../../../data/03_neurodegenerative_diseases/several_traits/four_traits/MaxH/03_gc_corrected/maxh_gc_corrected.csv', sep = '\t')
