==> 04_psychometry/14_significant_loci_sh.R <==
# Aim of this script is to count loci significant both for the original trait and SGIT

trait_designation <- c('bip','mdd','scz','happiness')

library(data.table)

clump <- fread('../../../data/03_neurodegenerative_diseases/several_traits/four_traits/clumping_of_SH_and_orig_traits_partIV_5e-8.txt')

ind <- grep('SH', clump$traits) # select loci significant for SGIT
sh <- clump[ind, ]

significant <- lapply(trait_designation, function (x) grepl(paste0('\\b', x, '\\b'), sh$traits, perl = T))
count <- sapply(significant, table)
count
